refactor: move value-mapping to its own module
- reduce circularity
- should be removed completly in v5
