+++ packages/core/src/helpers/import.ts
@@ -3,7 +3,7 @@ import { parseImports } from '@tokey/imports-parser';
 import { Diagnostics } from '../diagnostics';
 import type { Imported } from '../features';
 import { Root, decl, Declaration, atRule, rule, Rule, AtRule } from 'postcss';
-import { rootValueMapping, valueMapping } from '../stylable-value-parsers';
+import { rootValueMapping, valueMapping } from '../deprecated/value-mapping';
 import { stripQuotation } from '../helpers/string';
 import { isCompRoot } from './selector';
 import type { ParsedValue } from '../types';
