+++ packages/core/src/pseudo-states.ts
@@ -14,7 +14,8 @@ import type { PseudoClass } from '@tokey/css-selector-parser';
 import { StateResult, systemValidators } from './state-validators';
 import type { StylableMeta } from './stylable-meta';
 import type { StylableResolver } from './stylable-resolver';
-import { MappedStates, valueMapping } from './stylable-value-parsers';
+import type { MappedStates } from './stylable-value-parsers';
+import { valueMapping } from './deprecated/value-mapping';
 import type { ParsedValue, StateParsedValue } from './types';
 import { CSSClass } from './features';
 import { stripQuotation } from './helpers/string';
