+++ packages/core/src/stylable-value-parsers.ts
@@ -8,6 +8,7 @@ import { getNamedArgs } from './helpers/value';
 import type { StateParsedValue } from './types';
 import type { SelectorNodes } from '@tokey/css-selector-parser';
 import { CSSClass } from './features';
+import { valueMapping } from './deprecated/value-mapping';
 
 export interface MappedStates {
     [s: string]: StateParsedValue | string | null;
@@ -30,25 +31,6 @@ export interface ExtendsValue {
 }
 
 export type ReportWarning = (message: string, options?: { word: string }) => void;
-
-export const rootValueMapping = {
-    vars: ':vars' as const,
-    import: ':import' as const,
-    stScope: 'st-scope' as const,
-    namespace: 'namespace' as const,
-};
-
-export const valueMapping = {
-    from: '-st-from' as const,
-    named: '-st-named' as const,
-    default: '-st-default' as const,
-    root: '-st-root' as const,
-    states: '-st-states' as const,
-    extends: '-st-extends' as const,
-    mixin: '-st-mixin' as const, // ToDo: change to STMixin.MixinType.ALL,
-    partialMixin: '-st-partial-mixin' as const, // ToDo: change to STMixin.MixinType.PARTIAL,
-    global: '-st-global' as const,
-};
 
 export const mixinDeclRegExp = new RegExp(`(${valueMapping.mixin})|(${valueMapping.partialMixin})`);
